+++ src/scripts/util.c
@@ -171,10 +171,6 @@ static const char *lua_err_to_str(int rc) {
             return "Syntax error during precompilation";
         case LUA_ERRMEM:
             return "Memory allocation error";
-        #if LUA_VERSION_NUM >= 503 && LUA_VERSION_NUM < 504
-        case LUA_ERRGCMM:
-            return "Error in garbage collector";
-        #endif
         case LUA_ERRFILE:
             return "Can not open or read script file";
         case LUA_ERRRUN: