+++ CHANGELOG.md
@@ -8,7 +8,7 @@ https://github.com/jcorporation/myMPD/
 
 ### Notes
 
-This is the first release that supports only MPD 0.23.5 and higher and Lua 5.4.x
+This is the first release that supports only MPD 0.23.5 and higher and Lua 5.4.x.
 
 ### Changelog