+++ CHANGELOG.md
@@ -8,7 +8,7 @@ https://github.com/jcorporation/myMPD/
 
 ### Notes
 
-This is the first release that supports only MPD 0.23.5 and higher and Lua 5.4.x.
+This is the first release that supports only MPD 0.23.5 and higher and Lua 5.4.x and higher.
 
 ### Changelog
 
@@ -16,6 +16,7 @@ This is the first release that supports only MPD 0.23.5 and higher and Lua 5.4.x
 - Feat: Implement merge sort for linked lists
 - Feat: Regularly save the myMPD state if myMPD is active
 - Upd: Bump requirement for MPD and Lua versions
+- Upd: Search and utf8 handling improvements
 
 ***